update: 二分探索をlong型に変更

+++ Library/Search/Search.cs
@@ -7,14 +7,14 @@ namespace Suzuryg.Competitive.Library.Search
     public class Search
     {
         // https://qiita.com/drken/items/97e37dd6143e33a64c8c
-        public static int BinarySearch<T>(IReadOnlyList<T> list, Func<IReadOnlyList<T>, int, bool> isOK)
+        public static long BinarySearch<T>(IReadOnlyList<T> list, Func<IReadOnlyList<T>, long, bool> isOK)
         {
-            int ng = -1; //「index = 0」が条件を満たすこともあるので、初期値は -1
-            int ok = list.Count; // 「index = a.size()-1」が条件を満たさないこともあるので、初期値は list.Count
+            long ng = -1; //「index = 0」が条件を満たすこともあるので、初期値は -1
+            long ok = list.Count; // 「index = a.size()-1」が条件を満たさないこともあるので、初期値は list.Count
             // ok と ng のどちらが大きいかわからないことを考慮
             while (System.Math.Abs(ok - ng) > 1)
             {
-                int mid = (ok + ng) / 2;
+                long mid = (ok + ng) / 2;
 
                 if (isOK(list, mid))
                 {
@@ -27,12 +27,12 @@ namespace Suzuryg.Competitive.Library.Search
             }
             return ok;
         }
-        public static int BinarySearch(int ng, int ok, Func<int, bool> isOK)
+        public static long BinarySearch(long ng, long ok, Func<long, bool> isOK)
         {
             // ok と ng のどちらが大きいかわからないことを考慮
             while (System.Math.Abs(ok - ng) > 1)
             {
-                int mid = (ok + ng) / 2;
+                long mid = (ok + ng) / 2;
 
                 if (isOK(mid))
                 {
